mode auth routes to namespace
